docs/resources/source-code/ISWC/src/Reporting/requirements.txt
azure-mgmt-datafactory==0.13.0
opencensus-ext-azure==1.0.2
numpy==1.22.1
pandas==1.3.5
paramiko==2.11.0
pyarrow==19.0.1
pysftp==0.2.9
pytest==3.8.1
requests==2.28.1
setuptools==58.0.4